requirements.txt
BeautifulSoup==3.2.1
Django==1.4.1
Markdown==2.3.1
MySQL-python==1.2.3c1
PIL==1.1.7
South==0.7.6
Unidecode==0.04.12
Whoosh==2.4.1
django-guardian==1.1.0.beta
django-haystack==1.2.7
django-userena==1.2.0
easy-thumbnails==1.2
markdown2==2.1.0
